# Conflicts:
#	main.cpp
#	src/Clases/Contratacion/Contratacion.cpp
#	src/Clases/Contratacion/Contratacion.h
#	src/Clases/Desarrollador/Desarrollador.cpp
#	src/Clases/Jugador/Jugador.cpp
#	src/Clases/Jugador/Jugador.h
#	src/Clases/Partida/Partida.cpp
#	src/Clases/Partida/Partida.h
#	src/Clases/PartidaIndividual/PartidaIndividual.cpp
#	src/Clases/Puntuacion/Puntuacion.cpp
#	src/Clases/Sistema.cpp
#	src/Clases/Suscripcion/Suscripcion.cpp
#	src/Clases/Videojuego/Videojuego.cpp

+++ src/Clases/Puntuacion/Puntuacion.cpp
@@ -8,16 +8,16 @@ using namespace std;
 class Puntuacion : public ICollectible
 {
 private:
-    int puntuacion;
+    double puntuacion;
     Videojuego *juego;
     Jugador *creador;
 
 public:
     Puntuacion(int, Videojuego *, Jugador *);
     ~Puntuacion();
-    void setPuntuacion(int);
+    void setPuntuacion(double);
     void setCreador(Jugador *);
-    int getPuntuacion();
+    double getPuntuacion();
     Jugador *getCreador();
 };
 
@@ -34,7 +34,7 @@ Puntuacion::~Puntuacion()
     this->creador = NULL;
 }
 
-void Puntuacion::setPuntuacion(int puntuacion)
+void Puntuacion::setPuntuacion(double puntuacion)
 {
     this->puntuacion = puntuacion;
 }
@@ -44,7 +44,7 @@ void Puntuacion::setCreador(Jugador *creador)
     this->creador = creador;
 }
 
-int Puntuacion::getPuntuacion()
+double Puntuacion::getPuntuacion()
 {
     return this->puntuacion;
 }
